requirements.txt
pandas==2.2.1
plotly==5.20.0
gunicorn
dash-tools
